# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lifehacker Rss Parser REST API",
        "description": "Simple Rest Api",
        "contact": {
            "name": "Oleksandr Petryna"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of posts",
                "description": "Returns list of posts",
                "operationId": "posts.index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Posts not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Store new post",
                "description": "Returns created post data",
                "operationId": "posts.store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Required field not found or Link field already save"
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post information",
                "description": "Returns post data",
                "operationId": "posts.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update existing post",
                "description": "Returns updated post data",
                "operationId": "posts.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Required field not found or Link field already save"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "posts.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StorePostRequest": {
                "title": "Store Post request",
                "description": "Store Post request body data",
                "required": [
                    "title",
                    "link",
                    "description",
                    "creator",
                    "pubdate'"
                ],
                "properties": {
                    "title": {
                        "title": "Post title",
                        "description": "Title of the new Post",
                        "type": "integer",
                        "format": "int64string",
                        "example": "New Title"
                    },
                    "link": {
                        "title": "Post link",
                        "description": "Link of the new Post",
                        "type": "integer",
                        "format": "string",
                        "example": "https://new.example.com/"
                    },
                    "description": {
                        "title": "Post description",
                        "description": "Description of the new Post",
                        "type": "integer",
                        "format": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras at mollis sem. Sed in tellus id erat imperdiet consequat in eget arcu. Nulla tempus eget massa at aliquam. Cras augue urna, euismod id erat at, malesuada varius ex. Aliquam ullamcorper augue sem, vel suscipit mi consequat vel. Nunc a velit ut libero blandit posuere ut vitae lacus. Curabitur eros augue, malesuada vitae congue at, vehicula non justo. Praesent fringilla vel neque quis auctor. Mauris ultrices rhoncus dictum. Integer sed nisl eget tortor suscipit convallis. Quisque eget quam lorem. Sed eleifend placerat massa, non fermentum quam fringilla eget. Proin feugiat bibendum condimentum. Cras id luctus massa. In ut nunc faucibus, vestibulum augue quis, efficitur eros."
                    },
                    "creator": {
                        "title": "Post creator",
                        "description": "Creator of the new Post",
                        "type": "integer",
                        "format": "string",
                        "example": "John Smith"
                    },
                    "pubdate": {
                        "title": "Post publication date",
                        "description": "Pubication Date of the new Post",
                        "type": "integer",
                        "format": "date-time",
                        "example": "2022-04-14 04:10:00"
                    }
                },
                "type": "object"
            },
            "UpdatePostRequest": {
                "title": "Update Post request",
                "description": "Update Post request body data",
                "required": [
                    "id",
                    "title",
                    "link",
                    "description",
                    "creator",
                    "pubdate'"
                ],
                "properties": {
                    "title": {
                        "title": "Post title",
                        "description": "Title of the updated Post",
                        "type": "integer",
                        "format": "int64string",
                        "example": "Update Title"
                    },
                    "link": {
                        "title": "Post link",
                        "description": "Link of the updated Post",
                        "type": "integer",
                        "format": "string",
                        "example": "https://update.example.com/"
                    },
                    "description": {
                        "title": "Post description",
                        "description": "Description of the updated Post",
                        "type": "integer",
                        "format": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras at mollis sem. Sed in tellus id erat imperdiet consequat in eget arcu. Nulla tempus eget massa at aliquam. Cras augue urna, euismod id erat at, malesuada varius ex. Aliquam ullamcorper augue sem, vel suscipit mi consequat vel. Nunc a velit ut libero blandit posuere ut vitae lacus. Curabitur eros augue, malesuada vitae congue at, vehicula non justo. Praesent fringilla vel neque quis auctor. Mauris ultrices rhoncus dictum. Integer sed nisl eget tortor suscipit convallis. Quisque eget quam lorem. Sed eleifend placerat massa, non fermentum quam fringilla eget. Proin feugiat bibendum condimentum. Cras id luctus massa. In ut nunc faucibus, vestibulum augue quis, efficitur eros."
                    },
                    "creator": {
                        "title": "Post creator",
                        "description": "Creator of the updated Post",
                        "type": "integer",
                        "format": "string",
                        "example": "John Smith"
                    },
                    "pubdate": {
                        "title": "Post publication date",
                        "description": "Pubication Date of the updated Post",
                        "type": "integer",
                        "format": "date-time",
                        "example": "2022-04-14 04:10:00"
                    }
                },
                "type": "object"
            },
            "Post": {
                "title": "Post model",
                "description": "Post model",
                "required": [
                    "title",
                    "link",
                    "description",
                    "creator",
                    "pubdate"
                ],
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "title": "Post title",
                        "description": "Post title",
                        "type": "integer",
                        "format": "int64string",
                        "example": "New Title"
                    },
                    "link": {
                        "title": "Post link",
                        "description": "Post link",
                        "type": "integer",
                        "format": "string",
                        "example": "https://google.com/"
                    },
                    "description": {
                        "title": "Post description",
                        "description": "Post description",
                        "type": "integer",
                        "format": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras at mollis sem. Sed in tellus id erat imperdiet consequat in eget arcu. Nulla tempus eget massa at aliquam. Cras augue urna, euismod id erat at, malesuada varius ex. Aliquam ullamcorper augue sem, vel suscipit mi consequat vel. Nunc a velit ut libero blandit posuere ut vitae lacus. Curabitur eros augue, malesuada vitae congue at, vehicula non justo. Praesent fringilla vel neque quis auctor. Mauris ultrices rhoncus dictum. Integer sed nisl eget tortor suscipit convallis. Quisque eget quam lorem. Sed eleifend placerat massa, non fermentum quam fringilla eget. Proin feugiat bibendum condimentum. Cras id luctus massa. In ut nunc faucibus, vestibulum augue quis, efficitur eros."
                    },
                    "creator": {
                        "title": "Post creator",
                        "description": "Post creator",
                        "type": "integer",
                        "format": "string",
                        "example": "John Smith"
                    },
                    "pubdate": {
                        "title": "Post publication date",
                        "description": "Post publication date",
                        "type": "integer",
                        "format": "date-time",
                        "example": "2022-04-14 04:10:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Posts",
            "description": "First Project on Laravel with REST API"
        }
    ]
}